requirements.txt
absl-py==1.3.0
aiofile==3.8.5
asgiref==3.4.1
astunparse==1.6.3
autopep8==1.6.0
beautifulsoup4==4.9.3
cachetools==5.2.0
caio==0.9.12
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
cryptography==37.0.2
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==1.0.0
dj-iamport==0.0.1
Django==3.2.6
django-allauth==0.50.0
django-classy-tags==3.0.1
django-cors-headers==3.13.0
django-filter==23.2
django-heroku==0.3.1
django-seed==0.3.1
django-tagging==0.5.0
django-taggit==3.0.0
django-taggit-templatetags2==1.6.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
easydict==1.10
ez-setup==0.9
Faker==15.0.0
filelock==3.6.0
flatbuffers==22.10.26
gast==0.4.0
google-api-core==2.11.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-speech==2.19.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==20.1.0
h5py==3.7.0
idna==3.2
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.23.4
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==9.1.0
pipenv==2022.3.28
platformdirs==2.5.1
proto-plus==1.22.2
protobuf==4.22.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
stripe==4.0.2
tagging==0.2.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
toml==0.10.2
toposort==1.7
typing_extensions==4.4.0
tzdata==2022.1
urllib3==1.26.6
virtualenv==20.14.0
virtualenv-clone==0.5.7
websockets==11.0.2
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
